box2d update

+++ TankWars/core/src/main/java/com/mygdx/game/model/Position.java
@@ -4,27 +4,27 @@ package com.mygdx.game.model;
  * Created by marianarale on 2018-04-24.
  */
 public class Position {
-    private double x;
-    private double y;
+    private float x;
+    private float y;
 
-    public Position(double x, double y){
+    public Position(float x, float y){
         this.x = x;
         this.y = y;
     }
 
-    public double getX() {
+    public float getX() {
         return x;
     }
 
-    public double getY() {
+    public float getY() {
         return y;
     }
 
-    public void setX(double x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public void setY(double y) {
+    public void setY(float y) {
         this.y = y;
     }
 }
